Hide internal function normalized_vector
