+++ luda-adventure/src/app/game/types/game_object/types/collider/rigid_body/mod.rs
@@ -1,11 +1,9 @@
 mod circle;
 mod collide_circle_to_polygon;
 mod collide_polygon_to_circle;
-mod normalized_vector;
 mod rigid_body;
 
 pub use circle::*;
 pub use collide_circle_to_polygon::*;
 pub use collide_polygon_to_circle::*;
-pub use normalized_vector::*;
 pub use rigid_body::*;
